+++ language-server/server.go
@@ -17,6 +17,7 @@ package languageserver
 
 import (
 	"fmt"
+	"github.com/daveshanley/vacuum/loader"
 	"log/slog"
 	"os"
 	"path/filepath"
@@ -280,11 +281,21 @@ func (s *ServerState) Run() error {
 func (s *ServerState) runDiagnostic(doc *Document, notify glsp.NotifyFunc) {
 	// Copy document data while holding read lock to avoid data race
 	doc.mu.RLock()
-	content := doc.Content
 	uri := doc.URI
-	doc.mu.RUnlock()
+	currentPath := strings.TrimPrefix(uri, "file://")
 
-	specFileName := strings.TrimPrefix(uri, "file://")
+	var specFileName string
+	var content []byte
+	if s.lintRequest.MainSpecPath != "" {
+		specFileName = s.lintRequest.MainSpecPath
+		reportLoadResult, _ := loader.LoadFileAsReportOrSpecWithClient(specFileName, nil)
+		content = reportLoadResult.SpecBytes
+	} else {
+		specFileName = currentPath
+		content = []byte(doc.Content)
+	}
+
+	doc.mu.RUnlock()
 
 	// Copy config data while holding config lock to avoid data race
 	s.configMu.RLock()
@@ -299,7 +310,7 @@ func (s *ServerState) runDiagnostic(doc *Document, notify glsp.NotifyFunc) {
 	// Build the rule execution config with copied values
 	ruleExec := &motor.RuleSetExecution{
 		RuleSet:                         s.lintRequest.SelectedRS,
-		Spec:                            []byte(content),
+		Spec:                            content,
 		SpecFileName:                    specFileName,
 		Timeout:                         time.Duration(s.lintRequest.TimeoutFlag) * time.Second,
 		NodeLookupTimeout:               time.Duration(s.lintRequest.LookupTimeoutFlag) * time.Millisecond,
@@ -337,7 +348,7 @@ func (s *ServerState) runDiagnostic(doc *Document, notify glsp.NotifyFunc) {
 		}
 		filteredResults := utils.FilterIgnoredResultsWithOptions(result.Results, ignoredResults, ignoreOptions)
 		result.Results = filteredResults
-		diagnostics := ConvertResultsIntoDiagnostics(result)
+		diagnostics := ConvertResultsIntoDiagnostics(result, s.lintRequest.MainSpecPath, currentPath)
 
 		notify(protocol.ServerTextDocumentPublishDiagnostics, protocol.PublishDiagnosticsParams{
 			URI:         uri,
@@ -346,12 +357,19 @@ func (s *ServerState) runDiagnostic(doc *Document, notify glsp.NotifyFunc) {
 	}()
 }
 
-func ConvertResultsIntoDiagnostics(result *motor.RuleSetExecutionResult) []protocol.Diagnostic {
+func ConvertResultsIntoDiagnostics(result *motor.RuleSetExecutionResult, mainSpecPath string, currentPath string) []protocol.Diagnostic {
 	diagnostics := []protocol.Diagnostic{}
 
 	for _, vacuumResult := range result.Results {
-		diagnostics = append(diagnostics, ConvertResultIntoDiagnostic(&vacuumResult))
-
+		var source string
+		if vacuumResult.Origin == nil {
+			source = mainSpecPath
+		} else {
+			source = vacuumResult.Origin.AbsoluteLocation
+		}
+		if mainSpecPath == "" || filepath.Clean(source) == filepath.Clean(currentPath) {
+			diagnostics = append(diagnostics, ConvertResultIntoDiagnostic(&vacuumResult))
+		}
 	}
 	return diagnostics
 }
